Move null release test from FADTest to UpdateAppClientTest
