+++ firebase-app-distribution/src/test/java/com/google/firebase/appdistribution/FirebaseAppDistributionTest.java
@@ -17,7 +17,6 @@ package com.google.firebase.appdistribution;
 import static com.google.firebase.appdistribution.FirebaseAppDistributionException.Status.AUTHENTICATION_CANCELED;
 import static com.google.firebase.appdistribution.FirebaseAppDistributionException.Status.AUTHENTICATION_FAILURE;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
@@ -392,17 +391,5 @@ public class FirebaseAppDistributionTest {
     verify(mockSignInStorage).setSignInStatus(true);
     firebaseAppDistribution.signOutTester();
     verify(mockSignInStorage).setSignInStatus(false);
-  }
-
-  @Test
-  public void updateAppTask_whenNoReleaseAvailable_throwsError() throws Exception {
-    when(mockSignInStorage.getSignInStatus()).thenReturn(true);
-    UpdateTask updateTask = firebaseAppDistribution.updateApp();
-    assertFalse(updateTask.isSuccessful());
-    assertTrue(updateTask.getException() instanceof FirebaseAppDistributionException);
-    FirebaseAppDistributionException ex =
-        (FirebaseAppDistributionException) updateTask.getException();
-    assertEquals(FirebaseAppDistributionException.Status.UPDATE_NOT_AVAILABLE, ex.getErrorCode());
-    assertEquals(Constants.ErrorMessages.NOT_FOUND_ERROR, ex.getMessage());
   }
 }
